type.Genotype --> genotype.ObservedGenotype
