+++ microhapulator/genotype.py
@@ -9,6 +9,7 @@
 
 from collections import defaultdict
 from io import StringIO
+import json
 import microhapdb
 import microhapulator
 
@@ -69,3 +70,46 @@ class SimulatedGenotype(object):
         for line in self.bedstream:
             print(line, file=out)
         return out.getvalue()
+
+
+class ObservedGenotype(object):
+    def __init__(self, filename=None):
+        self.data = dict()
+        if filename:
+            with microhapulator.open(filename, 'r') as fh:
+                self.data = json.load(fh)
+
+    def record_coverage(self, locusid, cov_by_pos, ndiscarded=0):
+        self.data[locusid] = {
+            'mean_coverage': round(sum(cov_by_pos) / len(cov_by_pos), 1),
+            'min_coverage': min(cov_by_pos),
+            'max_coverage': max(cov_by_pos),
+            'num_discarded_reads': ndiscarded,
+            'allele_counts': dict(),
+        }
+
+    def record_allele(self, locusid, allele, count):
+        self.data[locusid]['allele_counts'][allele] = count
+
+    def infer(self):
+        for locusid, locusdata in self.data.items():
+            allelecounts = locusdata['allele_counts']
+            avgcount = sum(allelecounts.values()) / len(allelecounts.values())
+            gt = set()
+            for allele, count in allelecounts.items():
+                if count * 4 < avgcount:
+                    continue
+                gt.add(allele)
+            self.data[locusid]['genotype'] = sorted(gt)
+
+    def dump(self, file=None):
+        if file is None:
+            return json.dumps(self.data, indent=4, sort_keys=True)
+        else:
+            return json.dump(self.data, file, indent=4, sort_keys=True)
+
+    def alleles(self):
+        a = dict()
+        for locusid in sorted(self.data):
+            a[locusid] = ':'.join(self.data[locusid]['genotype'])
+        return a
